requirements.txt
fastapi
pydantic[email]
uvicorn
sqlalchemy
python-dotenv
alembic
aiosqlite
python-multipart
pillow
aiosmtplib